Move files to metadata/ folder
