+++ scripts/apps/authoring/media/metadata/MediaMetadataViewDirective.ts
@@ -5,7 +5,7 @@ export default function MediaMetadataViewDirective(deployConfig) {
             item: '=',
             showAltText: '@',
         },
-        template: require('./views/media-metadata-view-directive.html'),
+        template: require('../views/media-metadata-view-directive.html'),
         link: (scope) => {
             scope.validator = deployConfig.getSync('validator_media_metadata');
         },
